Derive a table's row type on AdoSchema, as JdbcSchema does
AdoTable derived its own proto row type, reading AdoDatabaseMetadata.GetFields
off a data source it held. That was defensible on its own terms and it was not
JdbcSchema's design, which is the one this adapter follows until something makes
it impossible.

AdoSchema now carries getRelDataType's two overloads: the three-argument one
acquiring the column metadata, the four-argument one deriving the type from it,
and sqlType as a private static beside them. AdoTable holds its AdoSchema the
way JdbcTable holds its jdbcSchema, its memoized supplier delegates through
SupplyProto as JdbcTable.supplyProto does, and DataSource, Convention and
Dialect read through the schema rather than through fields of their own.

The forced deviation is one parameter. JDBC acquires column metadata as
DatabaseMetaData off a connection the schema opens; ADO.NET has no equivalent --
GetSchema differs per provider and ODBC and OleDb cannot reliably say what they
are -- so this adapter owns the SPI as AdoDatabaseMetadata, and it stands where
DatabaseMetaData stands in the signature. The acquiring overload reads it from
the data source where upstream opens a connection. The comment says so at the
site, and no longer explains why the method is somewhere else.
